Fix FMOD API, narrowing, memory, and typedef issues
- Replaced deprecated/non-existent FMOD 3.75 API calls and fixed callback signatures/usage
- Added explicit default constructors to GFX vertex structs for STL compatibility
- Fixed narrowing conversion warnings with explicit casts
- Changed raw buffer allocation in SceneDraw to use ::operator new/delete
- Updated RiverBuilder to use emplace_back() for deque
- Corrected unordered_map::resize() to reserve() in TerrainInternal
- Moved CObjVisObj typedef to protected for derived class access
- Updated project files for BINK SDK paths and formatting
- Updated binary outputs and logs to reflect these changes

+++ Sources/src/Scene/Cursor.cpp
@@ -207,7 +207,7 @@ void CCursor::LockPos( bool bLock )
 		{
 			if ( bAcquired ) 
 			{
-				const RECT rcClip = { vPos.x, vPos.y, vPos.x, vPos.y };
+				const RECT rcClip = { static_cast<LONG>(vPos.x), static_cast<LONG>(vPos.y), static_cast<LONG>(vPos.x), static_cast<LONG>(vPos.y) };
 				::ClipCursor( &rcClip );
 			}
 		}
